ankimaker/requirements.txt
anki==25.9.2
certifi==2025.11.12
charset-normalizer==3.4.4
decorator==5.2.1
idna==3.11
Markdown==3.10
orjson==3.11.5
protobuf==6.33.2
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
